requirements.txt
GitPython==0.3.2.RC1
Django
